Use last incremental from index properties #660
