+++ minecode_pipelines/pipes/maven.py
@@ -7,26 +7,29 @@
 # See https://aboutcode.org for more information about nexB OSS projects.
 #
 
+from collections import namedtuple
 from itertools import chain
 import os
 import gzip
 import io
 import logging
-import arrow
-import javaproperties
-from packageurl import PackageURL
 
 from dateutil import tz
-from minecode_pipelines.pipes import java_stream
-from collections import namedtuple
-from scanpipe.pipes.fetch import fetch_http
-from scanpipe.pipes import federatedcode
-from minecode_pipeline.pipes import write_purls_to_repo
 from jawa.util.utf import decode_modified_utf8
-from packagedcode.maven import get_urls
+import arrow
+import javaproperties
+
 from packagedcode.maven import build_filename
 from packagedcode.maven import build_url
+from packagedcode.maven import get_urls
 from packagedcode.models import PackageData
+from packageurl import PackageURL
+from scanpipe.pipes.fetch import fetch_http
+from scanpipe.pipes import federatedcode
+
+from minecode_pipelines import miners
+from minecode_pipelines import pipes
+from minecode_pipelines.pipes import java_stream
 
 
 logger = logging.getLogger(__name__)
@@ -44,8 +47,9 @@ if TRACE:
 
 MAVEN_BASE_URL = "https://repo1.maven.org/maven2"
 MAVEN_INDEX_URL = "https://repo1.maven.org/maven2/.index/nexus-maven-repository-index.gz"
-MAVEN_INDEX_INCREMENT_BASE_URL = f"https://repo1.maven.org/maven2/.index/nexus-maven-repository-index.{index}.gz"
+MAVEN_INDEX_INCREMENT_BASE_URL = "https://repo1.maven.org/maven2/.index/nexus-maven-repository-index.{index}.gz"
 MAVEN_INDEX_PROPERTIES_URL = "https://repo1.maven.org/maven2/.index/nexus-maven-repository-index.properties"
+MAVEN_SETTINGS_PATH = "minecode_checkpoints/maven.json"
 
 
 def is_worthy_artifact(artifact):
@@ -577,6 +581,14 @@ class MavenNexusCollector:
     WARNING: Processing is rather long: a full index is ~600MB.
     """
 
+    def __init__(self, index_properties_location=None):
+        if index_properties_location:
+            content = index_properties_location
+        else:
+            content = self.fetch_index_properties()
+        with open(content) as config_file:
+            self.index_properties = javaproperties.load(config_file) or {}
+
     def fetch_index(self, uri=MAVEN_INDEX_URL):
         """
         Return a temporary location where the maven index was saved.
@@ -595,11 +607,11 @@ class MavenNexusCollector:
         """
         Yield maven index increments
         """
-        content = self.fetch_index_properties()
-        with open(content) as config_file:
-            properties = javaproperties.load(config_file) or {}
-
-        for key, increment_index in properties.items():
+        # in this context, last serial means last incremental
+        last_incremental = pipes.fetch_last_serial_mined(settings_path=MAVEN_SETTINGS_PATH)
+        for key, increment_index in self.index_properties.items():
+            if increment_index <= last_incremental:
+                continue
             if key.startswith("nexus.index.incremental"):
                 index_increment_url = MAVEN_INDEX_INCREMENT_BASE_URL.format(index=increment_index)
                 index_increment = fetch_http(index_increment_url)
@@ -676,9 +688,10 @@ class MavenNexusCollector:
         for index_increment in self.fetch_index_increments():
             return self._get_packages(content=index_increment)
 
-    def get_packages(self, content=None, increments=False):
+    def get_packages(self, content=None):
         """Yield Package objects from maven index"""
-        if increments:
+        last_incremental = pipes.fetch_last_serial_mined(settings_path=MAVEN_SETTINGS_PATH)
+        if last_incremental:
             packages = chain(self._get_packages_from_index_increments())
         else:
             if content:
@@ -698,7 +711,6 @@ def collect_packages_from_maven(commits_per_push=10, logger=None):
         if not prev_purl:
             prev_purl = current_purl
         elif prev_purl != current_purl:
-            # check out repo
             repo_url, _ = federatedcode.get_package_repository(
                 project_purl=prev_purl,
                 logger=logger
@@ -710,16 +722,18 @@ def collect_packages_from_maven(commits_per_push=10, logger=None):
 
             push_commit = not bool(i % commits_per_push)
             # save purls to yaml
-            write_purls_to_repo(
+            miners.write_purls_to_repo(
                 repo=repo,
                 package=prev_purl,
                 packages=current_packages,
                 push_commit=push_commit
             )
 
-            # delete local clone
             federatedcode.delete_local_clone(repo)
 
             current_packages = []
             prev_purl = current_purl
         current_packages.append(package)
+
+        last_incremental = maven_nexus_collector.index_properties.get("nexus.index.last-incremental")
+        pipes.update_last_serial_mined(last_serial=last_incremental, settings_path=MAVEN_SETTINGS_PATH)
